refactor: moving from vue to react
