+++ astro.config.mjs
@@ -1,10 +1,10 @@
 // @ts-check
 import { defineConfig } from "astro/config";
-import AstroVue from "@astrojs/vue";
+import AstroReact from "@astrojs/react";
 
 // https://astro.build/config
 export default defineConfig({
-  integrations: [AstroVue({})],
+  integrations: [AstroReact({})],
   redirects: {
     discord: "https://discord.com/invite/k8xj8d75f6",
   },
