+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "dev": "astro dev",
     "start": "astro dev",
-    "lint:css": "stylelint --formatter=verbose **/*.css **/*.vue",
+    "lint:css": "stylelint \"**/*.css\" --formatter=verbose ",
     "lint:etc": "prettier -c .",
     "lint": "npm run lint:css && npm run lint:etc",
     "lint:fix": "npm run lint:fix:css && npm run lint:fix:etc",
@@ -18,20 +18,23 @@
   },
   "dependencies": {
     "@astrojs/check": "^0.9.3",
-    "@astrojs/vue": "^4.5.1",
+    "@astrojs/react": "^3.6.2",
     "astro": "^4.15.10",
+    "clsx": "^2.1.1",
     "dayjs": "^1.11.13",
     "leaflet": "^1.9.4",
     "maplibre-gl": "^4.7.1",
-    "photoswipe": "^5.4.4",
-    "vue": "^3.5.10"
+    "photoswipe": "^5.4.4"
   },
   "devDependencies": {
     "@types/leaflet": "^1.9.12",
+    "@types/react": "^18.3.10",
+    "@vitejs/plugin-react": "^4.3.2",
     "@vitest/coverage-v8": "^2.1.1",
     "jsdom": "^25.0.1",
     "marked": "^14.1.2",
     "prettier": "^3.3.3",
+    "react": "^18.3.1",
     "stylelint": "^16.9.0",
     "stylelint-config-recommended-vue": "^1.5.0",
     "stylelint-config-standard": "^36.0.1",
