--- src/components/SiteFooter.vue
@@ -1,8 +1,0 @@
-<template>
-  <footer class="site-footer">
-    <p><span class="capitalize">Copyright</span> © {{ new Date().getFullYear() }} Osaka/Kyoto Web Designers and web developers meetup - <a href="/coc">Code of Conduct</a></p>
-  </footer>
-</template>
-<script setup lang="ts">
-import "./SiteFooter.css";
-</script>
