+++ src/components/events/EventDateDisplay.css
@@ -2,6 +2,7 @@
   margin-top: 0.1rem;
   max-width: 100%;
   background-color: #e5e7eb;
+
   /* TailwindCSS Gray 200 */
   padding: 0.5rem;
   border-radius: 10px;
