+++ src/components/header/SocialMedia.css
@@ -9,5 +9,5 @@
 }
 
 .social-media svg:hover {
-  fill: blue;
+  fill: red;
 }
