+++ src/layouts/Layout.astro
@@ -1,15 +1,14 @@
 ---
-import SiteHeader from '../components/SiteHeader.vue';
-import SiteFooter from '../components/SiteFooter.vue';
+import { SiteHeader } from "../components/SiteHeader.tsx";
+import { SiteFooter } from "../components/SiteFooter.tsx";
 
-import "./style.css"
-
+import "./style.css";
 
 const title = "Osaka Kyoto Web Designers and Developers Meetup";
 const description =
   "OWDDM and KWDDM are volunteer non-profit groups that organizes Meetups in Kansai. We try to have 1 meetup per month and Region with one or two presenters talking about a topic related to either making things in the Web. Occasionally we hold other events such as Hanami’s, hikes and other get-togethers.";
-
 ---
+
 <html lang="en">
   <head>
     <meta charset="utf-8" />
@@ -17,9 +16,9 @@ const description =
     <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
     <meta name="name" content="Osaka Kyoto Web Designers and Developers Meetup" />
     <meta name="title" content={title} />
-    <meta name="description" content={description}>
-    <meta property="og:title" content={title}>
-    <meta property="og:description" content={description}>
+    <meta name="description" content={description} />
+    <meta property="og:title" content={title} />
+    <meta property="og:description" content={description} />
     <title>{title}</title>
   </head>
   <body>
