+++ src/pages/about.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../layouts/Layout.astro";
-import AboutPage from '../components/pages/About.vue'
+import { AboutPage } from "../components/pages/About.tsx";
 ---
 
 <Layout>
