+++ src/pages/coc-report.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../layouts/Layout.astro'
-import CocReportPage from '../components/pages/CocReport.vue'
+import { CocReportPage } from '../components/pages/CocReport.tsx'
 ---
 
 <Layout>
