+++ src/pages/coc.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../layouts/Layout.astro'
-import CocPage from '../components/pages/Coc.vue'
+import { CocPage } from '../components/pages/Coc.tsx'
 ---
 
 <Layout>
