+++ src/pages/events/[event_id].astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '../../layouts/Layout.astro'
 import { fetchEvents, transform } from '../../utils/events'
-import EventPage from '../../components/pages/Event.vue'
+import { EventPage } from '../../components/pages/Event.tsx'
 
 export async function getStaticPaths() {
   const events = transform(await fetchEvents());
