+++ src/utils/lang.ts
@@ -1,0 +1,16 @@
+const base = "https://owddm.com/public";
+
+export function getSitePath(path: string): string {
+  return `${base}${path}`;
+}
+
+/**
+ * Returns an array of unique items from an array of items
+ *
+ * @param items
+ * @returns array of unique items
+ */
+export const getUniqueItems = <T>(items: T[]) => {
+  let unique = items.reduce<T[]>((acc, currentValue) => (acc.includes(currentValue) ? acc : [...acc, currentValue]), []);
+  return unique;
+};
