+++ vitest.config.ts
@@ -1,10 +1,10 @@
 /// <reference types="vitest" />
 
 import { defineConfig } from "vite";
-import vue from "@vitejs/plugin-vue";
+import react from "@vitejs/plugin-react";
 
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [react()],
   test: {
     globals: true,
     environment: "jsdom",
